Refactor Redis cache configuration
